Clear re-calibrations on 'regenerateState'.  Add more properties and relationships to buildings.

+++ common/models/building.js
@@ -10,24 +10,39 @@ module.exports = function(Building) {
 	Building.regenerateState = function(id, cb) {
 		var app = Building.app;
 		var State = app.models.State;
+		var Recalibration = app.models.Recalibration;
 
-		// Delete all existing States for the building.
-		State.destroyAll({
-			building : id
-		}, function(error, info) {
-			if (error) {
-				cb("Failed deleting states.");
-				console.log(error);
-				return;				
+		// Delete all existing Recalibration points for the Building.
+		Recalibration.destroyAll({
+			buildingId : id
+		}, function(destroyRecalibrationError) {
+			if (destroyRecalibrationError) {
+				cb("Failed while destroying Recalibration's.");
+				console.log(destroyRecalibrationError);
+				return;
 			}
-			console.log('destroy result');
-			console.log(info);
-		});
 
-		console.log('called regenerateState');
-		console.log('id');
-		console.log(id);
-		cb('something');
+			// Delete all existing States for the building.
+			State.destroyAll({
+				buildingId : id
+			}, function(destroyStateError, info) {
+				if (destroyStateError) {
+					cb("Failed deleting states.");
+					console.log(destroyStateError);
+					return;
+				}
+
+				// Fetch readings.
+					// TODO: Fetch in pages?
+
+					// For each reading.
+						// If a State reading point or midnight point has passed.
+							// Record the State.
+
+						// Update the current State as appropriate.
+							// Create Recalibration points where appropriate.
+			});
+		});
 	};
 
 	Building.remoteMethod(
